added new internal rule allowing for unopposed reelection of SC chair/vice-chair
As per discussion in ISSUE #265, adding new rule.

New rule added as number 7.9 as renumbering the rules would be
problematic (cause other references to rules 7.6-7.8 to become wrong!)

Text proposed by Michael Robbins based on his assessment:

“””
I see from the minutes of the March 2016 SC that the following
resolution was passed:

> RESOLUTION-SC-2016-01:Resolution: vote: 7-0 in favour - Amend
Internal Rules to say that default tenure is 2 years unless others want
to stand for election
>
> To change the default tenure of the SC Chair and SC Vice-Chair from 1
year to 2 years.

I’m afraid this is ultra vires as the term of office is set down in the
Article of Association (Article 13.6) as 12 months. It cannot be
changed without changing the Articles. I’m not sure this is really
necessary though – the “long stop” on consecutive years served is four
years. If an SC Chair / Vice Chair wants to continue in office and
there is a unanimous view that they should do so, then it would make
sense for them to be put up unopposed each year and voted back in as a
matter of course. You could probably have an internal rule along the
following lines…
“””