Removed rejected internal rule 7.9 that got accidentally included in a PR
Internal Rule relating to re-election of SC Chair / Vice-Chair was rejected during the 2018 Board Meeting. Unfortunately, the commit with this Rule change was also rolled up in a PR for another rule change. This commit removes the Rule (7.9 as was) and renumbers the rules accordingly.

The text of the deleted rule was:

```
In the event that a sitting SC Chair and/or SC Vice-Chair expresses their intention to stand for re-election under [Rule 7.5](#rule-7.5) and no other candidate wishes to stand against him or her, he or she may be re-elected to the role unopposed, provided that such re-election does not breach the term limit of four consecutive years laid down in [Article 13.6]({{ "/articles/13-steering-committee.html#article-13.6" | prepend: site.baseurl }}) of the Articles of Association.
```

The original closed PR with this rule change is here: https://github.com/OpenWIS/openwis-documentation/pull/337